+++ src/main/java/com/allrounders/goalkeeper/repository/RankRepository.java
@@ -3,5 +3,5 @@ package com.allrounders.goalkeeper.repository;
 import com.allrounders.goalkeeper.domain.Ranks;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RankRepository extends JpaRepository<Ranks, Integer>, RankRepositoryCustom {
+public interface RankRepository extends JpaRepository<Ranks, Long>, RankRepositoryCustom {
 }
